Former-commit-id: fff55c67a9568b553e46b9f472b78ddb418348d6 [formerly e80deec190b239bd3ca7d96979a3e7fef08dbf08] [formerly 95fb4ea8841895e32f5a501a0a0a1894281726ef [formerly ecc4e46fcbe172a25d2d88fa4f5c2f63f1af65dd]]
Former-commit-id: ae5ab55c09948003a81f0b17ae2b07273c63283c [formerly c65cd10251811d6263e0127337d68560b29300ee]
Former-commit-id: 55b66c36aabfda955b8a17c4a7361bbd9cbbc7c3

+++ iomodels/crmengine/opportunities.py
@@ -1107,7 +1107,7 @@ class Opportunity(EndpointsModel):
                     need = request.need
                     )
         if request.amount_total:
-            opportunity.amount_total = int(request.amount_total)
+            opportunity.amount_total = float(request.amount_total)
         else:
             opportunity.amount_total = 0
         opportunity_key = opportunity.put_async()
